+++ bin/check-es-cluster-health.rb
@@ -94,8 +94,9 @@ class ESClusterHealth < Sensu::Plugin::Check::CLI
 
   option :alert_status,
          description: 'Only alert when status matches given (red/yellow) option (or green)',
-         long: '--alert_status STATUS',
-         default: ''
+         long: '--alert-status STATUS',
+         default: '',
+         in: %w(RED YELLOW GREEN)
 
   def run
     options = {}
@@ -111,13 +112,13 @@ class ESClusterHealth < Sensu::Plugin::Check::CLI
     health = client.cluster.health options
     case health['status']
     when 'yellow'
-      if !options[:alert_status].downcase.include? 'red'
+      if !options[:alert_status] == 'RED'
         warning 'Cluster state is Yellow'
       else
         ok 'Not alerting on yellow'
       end
     when 'red'
-      if !options[:alert_status].downcase.include? 'yellow'
+      if !options[:alert_status] == 'YELLOW'
         critical 'Cluster state is Red'
       else
         ok 'Not alerting on red'